Replacing string.replace with string.slice
The current implementation of handleLanguageSubpath, which uses
string.replace, does not replace only the first occurrence of the
locale, but all matches.  So for the 'de' locale a URL like
'/de/page/delicious' would be truncated to '/pagelicious', because all
instances of '/de' would be replaced (both the one at the beginning
and the one in '/delicious').

This replaces string.replace with string.slice and removes only the
first occurrence of the locale in the URL.