fix: Do not call next() after redirect (#308)
* Do not call next() after redirect

If we perform a redirect in lngPathDetector (using localeSubpaths
option), we do not want any subsequent processing of routes or
middleware. Thus, next() should not be called in our middleware if we
redirect. This change ensures that we do not process any further
middleware or routes if we redirect in lngPathDetector.

Appropriate changes in unit and e2e tests have been made to ensure no
regression occurs.

Additionally, documentation changes have been made to make it explicit
that we may perform redirects within our middleware and, thus, the
order in which you run our middleware is important.

Lastly, splitting up middleware into distinct atomic units of work.

* Change function signature and redirect in middleware

Change function signature of lngPathDetector and redirect in middleware

* Remove redirectRequired from lngPathDetectorConfig object

Removing redirectRequired, as it is not strictly necessary, since we can always detect if it's needed by comparing originalUrl with correctedUrl.

* Changes per code review comments