docs: Clarify config for custom interpolation suffix/prefix
It took me a while to figure out how to get a custom
interpolation prefix/suffix to work.

I think it's a little unintuitive that changing an option
that relates to how the text inside the translation files
is structured has a knock-on effect to how files are loaded.

After doing some digging, I came across #1090 that clarifies
the issue, but I feel it's worth putting this in the README.

This PR does that. It's not much, but I hope you find it useful.