fix: Fix types for appWithTranslation
Updated appWithTranslation generic type to fit what is sent in by the
consumers of the library.
Marks _nextI18Next in SSRConfig as possibly null, which it may be in the
case of not calling serverSideTranslations in the page.
Added a test for the updated type in the related testfile.
Removed a few redundant null checks.